+++ rds-20140815/alibabacloud_rds20140815/models.py
@@ -4597,12 +4597,7 @@ class CreateDBInstanceRequest(TeaModel):
         # > *   If the instance runs PostgreSQL or MariaDB, you must set this parameter to **VPC**.
         # > *   RDS instances that run SQL Server Basic and SQL Server Web can reside in the classic network and virtual private clouds (VPCs). If the instance runs other database engines, you must set this parameter to **VPC**.
         self.instance_network_type = instance_network_type
-        # Specifies whether to enable the I/O acceleration feature. Valid values:
-        # 
-        # *   1: The feature is enabled.
-        # *   0: The feature is disabled.
-        # 
-        # >  For more information about the I/O acceleration feature of general enhanced SSDs (ESSDs), see [Introduction](~~2527067~~).
+        # A reserved parameter.
         self.io_acceleration_enabled = io_acceleration_enabled
         # The billing method of the instance. Valid values:
         # 
@@ -5298,12 +5293,7 @@ class CreateDBInstanceShrinkRequest(TeaModel):
         # > *   If the instance runs PostgreSQL or MariaDB, you must set this parameter to **VPC**.
         # > *   RDS instances that run SQL Server Basic and SQL Server Web can reside in the classic network and virtual private clouds (VPCs). If the instance runs other database engines, you must set this parameter to **VPC**.
         self.instance_network_type = instance_network_type
-        # Specifies whether to enable the I/O acceleration feature. Valid values:
-        # 
-        # *   1: The feature is enabled.
-        # *   0: The feature is disabled.
-        # 
-        # >  For more information about the I/O acceleration feature of general enhanced SSDs (ESSDs), see [Introduction](~~2527067~~).
+        # A reserved parameter.
         self.io_acceleration_enabled = io_acceleration_enabled
         # The billing method of the instance. Valid values:
         # 
@@ -17368,6 +17358,12 @@ class DescribeBackupPolicyResponseBody(TeaModel):
         # *   **True**: Incremental backup is enabled.
         # *   **False**: Incremental backup is disabled.
         self.enable_increment_data_backup = enable_increment_data_backup
+        # Whether PITR recovery is enabled at any point in time (the upgraded version of the original log backup). Return value:
+        # 
+        # - True: Yes
+        # - False: no
+        # 
+        # > Only MySQL instances return this parameter.
         self.enable_pitr_protection = enable_pitr_protection
         # Indicates whether the log backup deletion feature is enabled. If the disk usage exceeds 80% or the remaining disk space is less than 5 GB on the instance, this feature deletes binary log files. Valid values:
         # 
@@ -17389,6 +17385,7 @@ class DescribeBackupPolicyResponseBody(TeaModel):
         self.log_backup_local_retention_number = log_backup_local_retention_number
         # The number of days for which log backup files are retained.
         self.log_backup_retention_period = log_backup_retention_period
+        # The number of days to restore at any point in time.
         self.pitr_retention_period = pitr_retention_period
         # The cycle based on which you want to perform a backup. Separate multiple values with commas (,). Valid values:
         # 
@@ -22629,9 +22626,9 @@ class DescribeDBInstanceAttributeResponseBodyItemsDBInstanceAttribute(TeaModel):
         # 
         # >  This parameter applies only to ApsaraDB RDS for PostgreSQL instances for which Babelfish is enabled. For more information, see [Introduction to Babelfish](~~428613~~).
         self.babelfish_config = babelfish_config
-        # An invalid parameter. You can ignore this parameter.
+        # A deprecated parameter. You do not need to specify this parameter.
         self.bpe_enabled = bpe_enabled
-        # An invalid parameter. You can ignore this parameter.
+        # An invalid parameter. You do not need to specify this parameter.
         self.bursting_enabled = bursting_enabled
         # The RDS edition. Valid values:
         # 
@@ -22642,7 +22639,7 @@ class DescribeDBInstanceAttributeResponseBodyItemsDBInstanceAttribute(TeaModel):
         # *   **Finance**: RDS Enterprise Edition
         # *   **Serverless_basic**: RDS Basic Edition for serverless instances
         self.category = category
-        # A reserved parameter.
+        # A reserved parameter. You do not need to specify this parameter.
         self.cold_data_enabled = cold_data_enabled
         # The character set collation of the instance.
         self.collation = collation
@@ -22787,6 +22784,7 @@ class DescribeDBInstanceAttributeResponseBodyItemsDBInstanceAttribute(TeaModel):
         # 
         # *   **Postpaid**: pay-as-you-go
         # *   **Prepaid**: subscription
+        # *   **SERVERLESS**\
         self.pay_type = pay_type
         # The port that is used to connect to the instance over an internal network.
         self.port = port
@@ -25835,7 +25833,7 @@ class DescribeDBInstanceNetInfoResponseBodyDBInstanceNetInfosDBInstanceNetInfoDB
         self.availability = availability
         # The instance ID.
         self.dbinstance_id = dbinstance_id
-        # The role of the instance. Valid values:
+        # The type of the instance. Valid values:
         # 
         # *   **Master**: primary instance
         # *   **Readonly**: read-only instance
@@ -25924,7 +25922,7 @@ class DescribeDBInstanceNetInfoResponseBodyDBInstanceNetInfosDBInstanceNetInfoSe
     ):
         # The name of the IP address whitelist.
         self.security_ipgroup_name = security_ipgroup_name
-        # The IP addresses that is contained in the IP address whitelist.
+        # The IP address in the whitelist.
         self.security_ips = security_ips
 
     def validate(self):
@@ -26007,23 +26005,23 @@ class DescribeDBInstanceNetInfoResponseBodyDBInstanceNetInfosDBInstanceNetInfo(T
     ):
         # The Tabular Data Stream (TDS) port of the instance for which Babelfish is enabled.
         # 
-        # > This parameter applies only to ApsaraDB RDS for PostgreSQL instances. For more information about Babelfish for ApsaraDB RDS for PostgreSQL, see [Introduction to Babelfish](~~428613~~).
+        # >  This parameter applies only to ApsaraDB RDS for PostgreSQL instances. For more information about Babelfish for ApsaraDB RDS for PostgreSQL, see [Introduction to Babelfish](~~428613~~).
         self.babelfish_port = babelfish_port
-        # The endpoints of the instance.
+        # The endpoint of the instance.
         self.connection_string = connection_string
-        # The types of the endpoints. Valid values:
+        # The type of the endpoint. Valid values:
         # 
-        # *   **Normal**: regular endpoints
-        # *   **ReadWriteSplitting**: read/write splitting endpoints
+        # *   **Normal**: a regular endpoint
+        # *   **ReadWriteSplitting**: a read/write splitting endpoint
         self.connection_string_type = connection_string_type
         # The information about the instance weight.
         # 
-        # > This parameter is returned only for instances that have read/write splitting endpoints.
+        # >  This parameter is returned only when the read/write splitting feature is enabled for the instance.
         self.dbinstance_weights = dbinstance_weights
         # The policy that is used to assign read weights. This parameter is returned only for a read/write splitting endpoint. Valid values:
         # 
-        # *   **Standard**: The system automatically assigns read weights to the instance and its read-only instances based on the specifications of these instances.
-        # *   **Custom**: You must manually assign read weights to the instance and its read-only instances.
+        # *   **Standard**: The system automatically allocates read weights to the instance and its read-only instances based on the specifications of the instances.
+        # *   **Custom**: You must manually allocate read weights to the instance and its read-only instances.
         self.distribution_type = distribution_type
         # The remaining validity period of the instance in the classic network in hybrid access mode. Unit: seconds.
         self.expired_time = expired_time
@@ -26033,34 +26031,34 @@ class DescribeDBInstanceNetInfoResponseBodyDBInstanceNetInfosDBInstanceNetInfo(T
         # 
         # *   Valid values when the instance resides in the classic network:
         # 
-        #     *   **Inner:**: internal network
-        #     *   **Public**: Internet
+        #     *   **Inner**\
+        #     *   **Public**\
         # 
         # *   Valid values when the instance resides in a virtual private cloud (VPC):
         # 
-        #     *   **Private**: internal network
-        #     *   **Public**: Internet
+        #     *   **Private**\
+        #     *   **Public**\
         self.iptype = iptype
         # The latency threshold. This parameter is returned only for a read/write splitting endpoint. Unit: seconds.
         # 
-        # > If the latency on a read-only instance exceeds the specified threshold, the system no longer forwards read requests to the read-only instance.
+        # >  If the latency on a read-only instance exceeds the specified threshold, ApsaraDB RDS no longer forwards read requests to the read-only instance.
         self.max_delay_time = max_delay_time
         # The PgBouncer port.
         # 
-        # > This parameter is returned only when PgBouncer is enabled for the instance that runs PostgreSQL.
+        # >  This parameter is returned only when PgBouncer is enabled for the instance that runs PostgreSQL.
         self.pgbouncer_port = pgbouncer_port
         # The port that is used to connect to the instance.
         self.port = port
-        # The details of the IP address whitelist.
+        # The IP addresses in the whitelist for the instance.
         self.security_ipgroups = security_ipgroups
         # Indicates whether the IP version can be updated. Valid values:
         # 
-        # *   **Enable**: The IP version can be updated.
-        # *   **Disabled**: The IP version cannot be updated.
+        # *   **Enable**\
+        # *   **Disabled**\
         # 
-        # > The IP version can be updated from IPv4 to IPv6.
+        # >  The IP version can be updated from IPv4 to IPv6.
         self.upgradeable = upgradeable
-        # The VPC ID.
+        # The VPC ID of the instance.
         self.vpcid = vpcid
         # The vSwitch ID.
         self.v_switch_id = v_switch_id
@@ -28348,8 +28346,10 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
         connection_string: str = None,
         create_time: str = None,
         dbinstance_class: str = None,
+        dbinstance_cpu_cores: str = None,
         dbinstance_description: str = None,
         dbinstance_id: str = None,
+        dbinstance_memory: str = None,
         dbinstance_net_type: str = None,
         dbinstance_status: str = None,
         dbinstance_storage_type: str = None,
@@ -28410,6 +28410,7 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
         # 
         # >  This parameter is returned only when the **InstanceLevel** parameter is set to **1**.
         self.category = category
+        # A reserved parameter.
         self.cold_data_enabled = cold_data_enabled
         # The connection mode of the instance. Valid values:
         # 
@@ -28422,10 +28423,12 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
         self.create_time = create_time
         # The instance type of the instance. For information, see [Primary ApsaraDB RDS instance types](~~26312~~).
         self.dbinstance_class = dbinstance_class
+        self.dbinstance_cpu_cores = dbinstance_cpu_cores
         # The instance description.
         self.dbinstance_description = dbinstance_description
         # The instance ID.
         self.dbinstance_id = dbinstance_id
+        self.dbinstance_memory = dbinstance_memory
         # The type of the network connection to the instance. Valid values:
         # 
         # *   **Internet**\
@@ -28488,6 +28491,10 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
         # *   **Classic**\
         # *   **VPC**\
         self.instance_network_type = instance_network_type
+        # Whether IO acceleration is enabled. The value has the following meanings:
+        # 
+        # - 1: enabled
+        # - 0: not enabled
         self.io_acceleration_enabled = io_acceleration_enabled
         # The lock mode of the instance. Valid values:
         # 
@@ -28571,10 +28578,14 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
             result['CreateTime'] = self.create_time
         if self.dbinstance_class is not None:
             result['DBInstanceClass'] = self.dbinstance_class
+        if self.dbinstance_cpu_cores is not None:
+            result['DBInstanceCpuCores'] = self.dbinstance_cpu_cores
         if self.dbinstance_description is not None:
             result['DBInstanceDescription'] = self.dbinstance_description
         if self.dbinstance_id is not None:
             result['DBInstanceId'] = self.dbinstance_id
+        if self.dbinstance_memory is not None:
+            result['DBInstanceMemory'] = self.dbinstance_memory
         if self.dbinstance_net_type is not None:
             result['DBInstanceNetType'] = self.dbinstance_net_type
         if self.dbinstance_status is not None:
@@ -28677,10 +28688,14 @@ class DescribeDBInstancesResponseBodyItemsDBInstance(TeaModel):
             self.create_time = m.get('CreateTime')
         if m.get('DBInstanceClass') is not None:
             self.dbinstance_class = m.get('DBInstanceClass')
+        if m.get('DBInstanceCpuCores') is not None:
+            self.dbinstance_cpu_cores = m.get('DBInstanceCpuCores')
         if m.get('DBInstanceDescription') is not None:
             self.dbinstance_description = m.get('DBInstanceDescription')
         if m.get('DBInstanceId') is not None:
             self.dbinstance_id = m.get('DBInstanceId')
+        if m.get('DBInstanceMemory') is not None:
+            self.dbinstance_memory = m.get('DBInstanceMemory')
         if m.get('DBInstanceNetType') is not None:
             self.dbinstance_net_type = m.get('DBInstanceNetType')
         if m.get('DBInstanceStatus') is not None:
@@ -48195,6 +48210,7 @@ class DescribeSecretsRequest(TeaModel):
         self.page_size = page_size
         # The region ID. You can call the DescribeDBInstanceAttribute operation to query the region ID.
         self.region_id = region_id
+        # The ID of the resource group where the instance is located.
         self.resource_group_id = resource_group_id
         self.resource_owner_account = resource_owner_account
         self.resource_owner_id = resource_owner_id
@@ -53506,150 +53522,6 @@ class GrantOperatorPermissionResponse(TeaModel):
         return self
 
 
-class ImportDatabaseBetweenInstancesRequest(TeaModel):
-    def __init__(
-        self,
-        dbinfo: str = None,
-        dbinstance_id: str = None,
-        owner_account: str = None,
-        owner_id: int = None,
-        resource_owner_account: str = None,
-        resource_owner_id: int = None,
-        source_dbinstance_id: str = None,
-    ):
-        # The names of the source and destination databases. The value of this parameter is a JSON string.
-        # 
-        # >  If the source instance runs SQL Server, the value of this parameter consists of one or more key-value pairs. In each key-value pair, the key specifies the name of the source database, and the value specifies the name of the destination database. A source database can have a different name from the destination database. For example, `{"DBNames":{"srcdb":"destdb","srcdb2":"destmydb2"}}` indicates that srcdb is migrated to destdb and srcdb2 is migrated to destmydb2. The names of source databases cannot be the same, and the names of destination databases cannot be the same.
-        self.dbinfo = dbinfo
-        # The ID of the instance.
-        self.dbinstance_id = dbinstance_id
-        self.owner_account = owner_account
-        self.owner_id = owner_id
-        self.resource_owner_account = resource_owner_account
-        self.resource_owner_id = resource_owner_id
-        # The ID of the source instance, which cannot be the same as the ID of the destination instance.
-        self.source_dbinstance_id = source_dbinstance_id
-
-    def validate(self):
-        pass
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.dbinfo is not None:
-            result['DBInfo'] = self.dbinfo
-        if self.dbinstance_id is not None:
-            result['DBInstanceId'] = self.dbinstance_id
-        if self.owner_account is not None:
-            result['OwnerAccount'] = self.owner_account
-        if self.owner_id is not None:
-            result['OwnerId'] = self.owner_id
-        if self.resource_owner_account is not None:
-            result['ResourceOwnerAccount'] = self.resource_owner_account
-        if self.resource_owner_id is not None:
-            result['ResourceOwnerId'] = self.resource_owner_id
-        if self.source_dbinstance_id is not None:
-            result['SourceDBInstanceId'] = self.source_dbinstance_id
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('DBInfo') is not None:
-            self.dbinfo = m.get('DBInfo')
-        if m.get('DBInstanceId') is not None:
-            self.dbinstance_id = m.get('DBInstanceId')
-        if m.get('OwnerAccount') is not None:
-            self.owner_account = m.get('OwnerAccount')
-        if m.get('OwnerId') is not None:
-            self.owner_id = m.get('OwnerId')
-        if m.get('ResourceOwnerAccount') is not None:
-            self.resource_owner_account = m.get('ResourceOwnerAccount')
-        if m.get('ResourceOwnerId') is not None:
-            self.resource_owner_id = m.get('ResourceOwnerId')
-        if m.get('SourceDBInstanceId') is not None:
-            self.source_dbinstance_id = m.get('SourceDBInstanceId')
-        return self
-
-
-class ImportDatabaseBetweenInstancesResponseBody(TeaModel):
-    def __init__(
-        self,
-        import_id: str = None,
-        request_id: str = None,
-    ):
-        # The ID of the import task.
-        self.import_id = import_id
-        # The ID of the request.
-        self.request_id = request_id
-
-    def validate(self):
-        pass
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.import_id is not None:
-            result['ImportId'] = self.import_id
-        if self.request_id is not None:
-            result['RequestId'] = self.request_id
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('ImportId') is not None:
-            self.import_id = m.get('ImportId')
-        if m.get('RequestId') is not None:
-            self.request_id = m.get('RequestId')
-        return self
-
-
-class ImportDatabaseBetweenInstancesResponse(TeaModel):
-    def __init__(
-        self,
-        headers: Dict[str, str] = None,
-        status_code: int = None,
-        body: ImportDatabaseBetweenInstancesResponseBody = None,
-    ):
-        self.headers = headers
-        self.status_code = status_code
-        self.body = body
-
-    def validate(self):
-        if self.body:
-            self.body.validate()
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.headers is not None:
-            result['headers'] = self.headers
-        if self.status_code is not None:
-            result['statusCode'] = self.status_code
-        if self.body is not None:
-            result['body'] = self.body.to_map()
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('headers') is not None:
-            self.headers = m.get('headers')
-        if m.get('statusCode') is not None:
-            self.status_code = m.get('statusCode')
-        if m.get('body') is not None:
-            temp_model = ImportDatabaseBetweenInstancesResponseBody()
-            self.body = temp_model.from_map(m['body'])
-        return self
-
-
 class ImportUserBackupFileRequest(TeaModel):
     def __init__(
         self,
@@ -55401,7 +55273,7 @@ class MigrateToOtherZoneRequest(TeaModel):
         # 
         # >  If you set this parameter to **ScheduleTime**, you must specify the **SwitchTime** parameter.
         self.effective_time = effective_time
-        # A reserved parameter. You do not need to specify this parameter.
+        # A reserved parameter.
         self.io_acceleration_enabled = io_acceleration_enabled
         # Specifies whether to change the specifications of the instance during the cross-zone migration. Valid values:
         # 
@@ -56663,23 +56535,25 @@ class ModifyCollationTimeZoneRequest(TeaModel):
         resource_owner_id: int = None,
         timezone: str = None,
     ):
-        # The character set collation of the instance. Valid values:
+        # The character set collation of the instance. By default, the system does not modify the character set collation of the instance. Valid values:
         # 
+        # *   **Chinese_PRC_CI_AS**\
+        # *   **Chinese_PRC_CS_AS**\
+        # *   **Chinese_PRC_BIN**\
         # *   **Latin1\_General_CI_AS**\
         # *   **Latin1\_General_CS_AS**\
         # *   **SQL_Latin1\_General_CP1\_CI_AS**\
         # *   **SQL_Latin1\_General_CP1\_CS_AS**\
-        # *   **Chinese_PRC_CS_AS**\
-        # *   **Chinese_PRC_BIN**\
-        # *   **Chinese_PRC_CI_AS**\
         # *   **Japanese_CI_AS**\
         # *   **Japanese_CS_AS**\
         # *   **Chinese_Taiwan_Stroke_CI_AS**\
         # *   **Chinese_Taiwan_Stroke_CS_AS**\
         # 
-        # By default, the system does not modify the character set collation of the instance.
+        # > 
         # 
-        # > You must specify one of the **Collation** and **Timezone** parameters.
+        # *   The default character set collation of the instance is **Chinese_PRC_CI_AS**.
+        # 
+        # *   You must specify one of the **Collation** and **Timezone** parameters.
         self.collation = collation
         # The instance ID.
         self.dbinstance_id = dbinstance_id
@@ -56688,7 +56562,11 @@ class ModifyCollationTimeZoneRequest(TeaModel):
         self.resource_owner_id = resource_owner_id
         # The time zone of the instance. By default, the system does not modify the time zone.
         # 
-        # > You must specify one of the **Collation** and **Timezone** parameters.
+        # > 
+        # 
+        # *   The default time zone of the instance is **China Standard Time**.
+        # 
+        # *   You must specify one of the **Collation** and **Timezone** parameters.
         self.timezone = timezone
 
     def validate(self):
@@ -59820,7 +59698,7 @@ class ModifyDBInstanceSSLRequest(TeaModel):
         # The method that is used to verify the identities of clients. This parameter is supported only when the instance runs PostgreSQL with cloud disks. In addition, this parameter is available only when the public key of the CA that issues client certificates is enabled. Valid values:
         # 
         # *   **cert**\
-        # *   **perfer**\
+        # *   **prefer**\
         # *   **verify-ca**\
         # *   **verify-full** (supported only when the instance runs PostgreSQL 12 or later)
         self.acl = acl
@@ -59829,6 +59707,10 @@ class ModifyDBInstanceSSLRequest(TeaModel):
         # *   **aliyun**: a cloud certificate
         # *   **custom**: a custom certificate
         self.catype = catype
+        # User-defined certificate. The custom certificate is in pfx format.
+        # 
+        # - Public address: `oss-<region ID>.aliyuncs.com:<Bucket name >:< certificate file name (with file suffix)>`
+        # - Intranet address: `oss-<region ID>-internal.aliyuncs.com:<Bucket name >:< certificate file name (with file suffix)>`
         self.certificate = certificate
         # The public key of the CA that issues client certificates. This parameter is supported only when the instance runs PostgreSQL with cloud disks. This parameter must be specified when ClientCAEbabled is set to **1**.
         self.client_cacert = client_cacert
@@ -59855,11 +59737,12 @@ class ModifyDBInstanceSSLRequest(TeaModel):
         self.force_encryption = force_encryption
         self.owner_account = owner_account
         self.owner_id = owner_id
+        # The password of the certificate.
         self.pass_word = pass_word
-        # The method that is used to verify the replication permission. This parameter is supported only when the instance runs PostgreSQL with cloud disks. In addition, this parameter is available only when the public key of the CA that issues client certificates is enabled. Valid values:
+        # The method that is used to verify the replication permissions. This parameter is supported only when the instance runs PostgreSQL with cloud disks. In addition, this parameter is available only when the public key of the CA that issues client certificates is enabled. Valid values:
         # 
         # *   **cert**\
-        # *   **perfer**\
+        # *   **prefer**\
         # *   **verify-ca**\
         # *   **verify-full** (supported only when the instance runs PostgreSQL 12 or later)
         self.replication_acl = replication_acl
@@ -68748,12 +68631,9 @@ class ResetAccountPasswordRequest(TeaModel):
         self.account_name = account_name
         # The new password.
         # 
-        # > 
-        # 
-        # *   The value must be 8 to 32 characters in length.
-        # 
-        # *   The value must contain at least three types of the following characters: uppercase letters, lowercase letters, digits, and special characters.
-        # *   The following special characters are supported: ! @ # $ & % ^ \* ( ) \_ + - =\
+        # > *   The value must be 8 to 32 characters in length.
+        # > *   The value must contain at least three types of the following characters: uppercase letters, lowercase letters, digits, and special characters.
+        # > *   The following special characters are supported: ! @ # $ & % ^ \* ( ) \_ + - =\
         self.account_password = account_password
         # The instance ID. You can call the DescribeDBInstances operation to query the instance ID.
         self.dbinstance_id = dbinstance_id